Remove references to long() type, inspired using “2to3”
* L suffix not needed since Python 2.2; dropped from Python 3

+++ src/AppletLoader.py
@@ -116,7 +116,7 @@ class AppletLoader:
             value = int(value, 0)
         except ValueError:
             try:
-                value = long(value, 0)
+                value = int(value, 0)
             except ValueError:
                 try:
                     value = float(value)
